third_party/WebKit/LayoutTests/svg/as-image/svgview-references-use-counters-expected.txt

PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is false
PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is false
PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is true
PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is false
PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is true
PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is true
PASS successfullyParsed is true

TEST COMPLETE
  
